New version of install script (has not been tested yet)

+++ scripts/install_utcoupe_setup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 ### The goal of this script is to install all UTCoupe specific packages to have a working setup.
-### This script is called automatically when you run a npm install.
+### When you just cloned the UTCoupe repository, you have to run this script first !
 
 function green_echo() {
 	echo -e "\033[32m$1\033[0m"
@@ -17,7 +17,7 @@ ARCH=$(uname -m)
 ### Install the linux packages
 function install_apt() {
 	green_echo "Install missing packages..."
-	sudo apt-get install git build-essential python cmake libboost-dev libsdl1.2-dev gcc-avr avrdude avr-libc
+	sudo apt-get install git build-essential python python-pip cmake libboost-dev libsdl1.2-dev gcc-avr avrdude avr-libc
 
 	# Check if it's a PC or a raspi
 	if [ "$ARCH" = "x86_64" ]; then
@@ -43,6 +43,19 @@ function install_apt() {
 	fi
 }
 
+### Install ros-desktop-full
+function install_ros() {
+	if [ "$(lsb_release -sc)" = "xenial" ] || [ "$(lsb_release -sc)" = "willy" ]; then
+		sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros-latest.list'
+		sudo apt-key adv --keyserver hkp://ha.pool.sks-keyservers.net:80 --recv-key 421C365BD9FF1F717815A3895523BAEEB01FA116
+		sudo apt-get update
+		sudo apt-get install ros-kinetic-desktop-full
+		sudo rosdep init
+		rosdep update
+	else
+		red_echo "Your OS is not Ubuntu Willy or Xenial, ROS will not been installed..."
+	fi
+}
 
 ### Setup the variable environment to taget the UTCoupe main folder
 function env_setup() {
@@ -78,49 +91,32 @@ function env_setup() {
 	# Create the utcoupe folder where log files are stored
 	if [ ! -d "/var/log/utcoupe" ]; then
 		sudo mkdir /var/log/utcoupe
+		sudo chown $USER:$USER /var/log/utcoupe
 	fi
 	# Untar all libraries
+	#TODO untar only if it does not exist
 	for f in $UTCOUPE_WORKSPACE/libs/*.tar; do tar -C $UTCOUPE_WORKSPACE/libs xzf $f; done
 	# "Install" Arduino libs
-	sudo ln -s $UTCOUPE_WORKSPACE/libs/arduino-1.0 /opt/
-	# Change the ownership of the utcoupe log folder
-	sudo chown $USER:$USER /var/log/utcoupe
-	# Install the hokuyo automatic startup script (only for raspberry pi zero)
-	if [ ! -f "/etc/init.d/utcoupe_hokuyo.sh" ] && [ "$ARCH" = "armv6l" ]; then
-		sudo install $UTCOUPE_WORKSPACE/scripts/utcoupe_hokuyo.sh /etc/init.d/
-		sudo update-rc.d utcoupe_hokuyo.sh defaults 99
+	if [ ! -L "/opt/arduino-1.0" ]; then
+		sudo ln -s $UTCOUPE_WORKSPACE/libs/arduino-1.0 /opt/
 	fi
+	# Add the Ethernet IP address of raspberry pi to have a shortcut
+	if ! grep "utcoupe" /etc/hosts > /dev/null; then
+	    sudo sh -c "echo '#UTCoupe raspberry pi Ethernet IP when connected on the UTC network\n172.18.159.254	utcoupe_rpi31\n172.18.161.161	utcoupe_rpi32\n172.18.161.162	utcoupe_rpi33' >> /etc/hosts"
+	fi
+	#TODO add ssh to the raspi (must be connected...)
 }
 
-### Compile and install the UTCoupe libraries
-
-# URG library for the hokuyo
-function compile_urg() {
-	cd $UTCOUPE_WORKSPACE/libs/urg-0.8.18
-	./configure && make && sudo make install
+### Then install the UTCoupe ROS workspace
+function install_ros_workspace() {
+	# Download the submodules code
+	git submodule update --init --recursive
+	# Install the UTCoupe ROS specific packages
+	#TODO use the requirements system
+	sudo pip install pyserial numpy scipy pymongo
 }
 
-# Archer driver for 5 GHz wifi
-function compile_archer() {
-	cd $UTCOUPE_WORKSPACE/libs/Archer_T1U_V1_150909/Driver
-	sudo make && sudo make install
-	#TODO add the ra0 interface in configuration files
-}
-
-### Then install the UTCoupe softwares
-
-# The pathfinding
-function compile_pathfinding() {
-	cd $UTCOUPE_WORKSPACE/pathfinding
-	./make.sh
-}
-
-# The hokuyo
-function compile_hokuyo() {
-	cd $UTCOUPE_WORKSPACE/hokuyo
-	./make.sh
-}
-
+### Main install_script function, ask the user to install each main components
 function launch_script() {
 
 	env_setup
@@ -129,31 +125,21 @@ function launch_script() {
 	read answer
 	if [ "$answer" = "" ] || [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
 		install_apt
+		if [ ! -f "/opt/ros" ]; then
+			printf "ROS has not been detected in /opt/ros, launch the installation process..."
+			install_ros
+		fi
 	fi
-	
-	printf "Compile archer library (mandatory for 5 GHz usb wifi key) ? [Y/n]?"
+	printf "Install UTCoupe ROS workspace (ROS must to be installed) ? [Y/n]?"
 	read answer
 	if [ "$answer" = "" ] || [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
-		compile_archer
-	fi
-	
-	printf "Compile pathfinding ? [Y/n]?"
-	read answer
-	if [ "$answer" = "" ] || [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
-		compile_pathfinding
-	fi
-	
-	printf "Compiled hokuyo (+ urg mandatory library) ? [Y/n]?"
-	read answer
-	if [ "$answer" = "" ] || [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
-		compile_urg
-		compile_hokuyo
+		install_ros_workspace
 	fi
 }
 
 # Verify that the script is launched from the right place
-if [ ! "${PWD##*/}" = "coupe17" ]; then
-	red_echo "You have to launch this script from UTCoupe main directory : ./script/${0##*/}"
+if [ ! "${PWD##*/}" = "coupe18" ]; then
+	red_echo "You have to launch this script from UTCoupe main directory : ./script/${0##*/} or to rename this folder in coupe18."
 	exit 1
 fi
